VST.ML/requirements.txt
absl-py==1.0.0
appdirs==1.4.4
asttokens==2.0.5
astunparse==1.6.3
audioread==2.1.9
backcall==0.2.0
black==22.1.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
decorator==5.1.1
executing==0.8.2
flatbuffers==2.0
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.10.1
ipython==8.0.1
jedi==0.18.1
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==13.0.0
librosa==0.9.0
llvmlite==0.38.0
Markdown==3.3.6
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
numba==0.55.1
numpy==1.21.5
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.0.1
pip==21.2.3
platformdirs==2.5.0
pooch==1.6.0
prompt-toolkit==3.0.27
protobuf==3.19.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
pydub==0.25.1
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.2
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
setuptools==57.4.0
six==1.16.0
SoundFile==0.10.3.post1
stack-data==0.1.4
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tomli==2.0.1
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.0.3
wheel==0.37.1
wrapt==1.13.3
zipp==3.7.0
